+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [UITEN-330](https://folio-org.atlassian.net/browse/UITEN-330) *BREAKING* Refactor locale settings handling.
 * [UITEN-326](https://folio-org.atlassian.net/browse/UITEN-326) Settings > Tenant > Add Tenant application icon.
 * [UITEN-328](https://folio-org.atlassian.net/browse/UITEN-328) Settings > Tenant > Service points: view and edit the new `defaultCheckInActionForUseAtLocation` field. Requires v3.5 of the `service-points` interface.
+* [UITEN-331](https://folio-org.atlassian.net/browse/UITEN-331) Settings > Tenant > Service points: default undefined value for `defaultCheckInActionForUseAtLocation` field.
 
 ## [10.0.0](https://github.com/folio-org/ui-tenant-settings/tree/v10.0.0)(2025-03-12)